backend/requirements.txt
dj-database-url
annotated-types==0.7.0
anyio==4.11.0
asgiref==3.10.0
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
cryptography==46.0.3
deprecation==2.1.0
Django==5.2.7
django-cors-headers==4.9.0
django-filter==24.3
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
ecdsa==0.19.1
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
hyperframe==6.1.0
idna==3.11
multidict==6.7.0
packaging==25.0
pillow==12.0.0
postgrest==2.23.2
propcache==0.4.1
psycopg2-binary==2.9.11
pyasn1==0.6.1
pycparser==2.23
pydantic==2.12.4
pydantic_core==2.41.5
PyJWT==2.10.1
python-decouple==3.8
python-jose==3.5.0
realtime==2.23.2
requests==2.32.5
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.3
storage3==2.23.2
StrEnum==0.4.15
supabase==2.23.2
supabase-auth==2.23.2
supabase-functions==2.23.2
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.5.0
websockets==15.0.1
yarl==1.22.0
gunicorn==23.0.0
dj-database-url==2.3.0

# Security packages
django-ratelimit==4.1.0
django-otp==1.5.4
pyotp==2.9.0
qrcode==8.0
python-magic==0.4.27
bleach==6.2.0

# Caching
django-redis==5.4.0
redis==5.2.0

# Monitoring & Logging
sentry-sdk==2.20.0
django-prometheus==2.3.1

# Testing
locust==2.32.3
factory-boy==3.3.1
faker==33.1.0

# Data encryption
django-encrypted-model-fields==0.6.5
